CRASH:Memory access violation in posix_acl_2_fsal_acl
Commit 11d0261ff9d15079a9d3f7d8c9ff39342d4dd01b causes
posix_acl_2_fsal_acl() to generate ACEs in the FSAL ACL that
correspond to an input POSIX ACL mask ACE. However, ace_count(),
which is used to calculate space for the output, wasn't modified
to provide space for the additional ACEs. For some POSIX ACLs
this can cause the output to overrun the allocated space
resulting in a crash.

STARTUP :MAJ :stack backtrace follows:
/lib64/libganesha_nfsd.so.5.5.3(+0x96700)[0x7f17ba175700]
/lib64/libganesha_nfsd.so.5.5.3(+0x66d90)[0x7f17ba145d90]
/lib64/libpthread.so.0(+0x12cf0)[0x7f17b823fcf0]
/lib64/libc.so.6(gsignal+0x10f)[0x7f17b7a94acf]
/lib64/libc.so.6(abort+0x127)[0x7f17b7a67ea5]
/lib64/libc.so.6(+0x8fcd7)[0x7f17b7ad5cd7]
/lib64/libc.so.6(+0x96fdc)[0x7f17b7adcfdc]
/lib64/libc.so.6(+0x9b1fc)[0x7f17b7ae11fc]
/lib64/libc.so.6(realloc+0x33b)[0x7f17b7ae23db]
/usr/lib64/ganesha/libfsalvfs.so(vfs_sub_getattrs+0x32b)
[0x7f17b244d930]
/usr/lib64/ganesha/libfsalvfs.so(+0xf10d)[0x7f17b244710d]
/usr/lib64/ganesha/libfsalvfs.so(vfs_getattr2+0x23d)
[0x7f17b244a6f7]
/lib64/libganesha_nfsd.so.5.5.3(+0x16b786)[0x7f17ba24a786]

Change-Id: I81efee93da9f4d7eadbb2e0b36aefdcaf36d3faa
Signed-off-by: Arnab Tah <atah@ddn.com>