nfs_worker: Destroy connection when rpcsec_gss_data request is dropped
In case of `RPCSEC_GSS_DATA` procedure, if Ganesha receives an
AUTH_OK response from ntirpc with instructions to discard the
request, the actual authentication did not complete, and ntirpc
itself also did not send any response to the client. In such a
scenario, in order to remain compliant both GSS_API RFC and
NFSV4.1 RFC, Ganesha should terminate the RPC connection. This
CL addresses it.

Change-Id: I1fc81fef67ebf1292f1e537c34a6a2560a72b62b
Signed-off-by: Dipit Grover <dipit@google.com>