FSAL: Fix some issues with resolving POSIX filesystems
Commit 6e18be9236b952fd1dbba80b4ef6819e3f2f9a2d accidentally
introduced a 1 second delay per export.

If stat() returns 0 then the initial loop iteration continues, causing
at least one call to nanosleep().  Instead, the loop should break on
success.

Why 1 second?  Arguments to CONF_ITEM_UI32() should look like:

  CONF_ITEM_UI32(_name_, _min_, _max_, _def_, _struct_, _mem_)

That is the 3rd argument should be the maximum and the 4th should be
the default.  These currently appear to be in the wrong order, since
the default is not within the range.  This is actually lucky because
with a timeout of 100ms we probably wouldn't have noticed the first
problem.  ;-)

Signed-off-by: Martin Schwenke <mschwenke@ddn.com>
Change-Id: If5e0ca60485a2e5f13d03fa5e2a3050ef6697eab