Idmapper: Limit max user-groups cache entries, and add a fifo queue
This change limits the max number of user-groups entries in the
uid2grp cache. The limit is configurable through the
DIRECTORY_SERVICES section in Ganesha config.

The cache eviction strategy is determined using the fifo queue.
The fifo queue order mimics the order of expiration time of the
cache entries, since expiration time is a linear function of
the insertion time --

expiration_time = insertion_time + cache_validity (constant)

The head of the queue therefore contains the entry with least
time-validity, and the tail of the queue contains the entry
with most time-validity. The eviction happens from the head,
and insertion happens at the tail.

Change-Id: I8dae2b3597bc6604328dd6699d24eb128544fb93
Signed-off-by: Dipit Grover <dipit@google.com>