CEPH: Fix up cmount_path
OOPS - cmount_path is NOT mandatory

If it isn't specified, then handle like the old days.

Support an export without cmount_path (so the cmount will be at fullpath)
with exports of a sub-directory of that, specifying that path as
cmount_path. The result, as long as all other parameters are the same, will
be sharing of the cephfs client.

Change-Id: Iefc3043eca5a4445931bc9c95b916f198d50510d
Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>