Issue in ganesha_stats json option part_3
Issue: ganesha_stats json iomon is showing the stats twice.
The output is having 2 entries for each export_id

Fix: While creating the dict for showing json format stats
for a export_id were appended twice. This was because the
append operation was done in the inner loop. So during read
and write ops the whole stats were getting appneded.
Moved the append operation outside the inner loop.

Change-Id: I8263257012ca981a52ec73d4e22fdcd313cb86fc
Signed-off-by: Trupti Shete <tshete@ddn.com>