ganesha_stats json issues part_1
Issue: If no command is passed to ganesha_stats json in that case
it gives error.

Fix: The default command is not set for json option in
ganesha_stats due to that it is failing if no command is passed.
To resolve this issue DEFAULT_COMMAND is set to global and same
is used in case of no json. So in future if there is any change in
default command it will be applied in both the cases.

Change-Id: Icb8d4f88a57543fc0bed9921d05314be82c34460
Signed-off-by: Trupti Shete <tshete@ddn.com>