bind_conn_to_session: Validate check/add connection to session output
Currently during bind_conn_to_session op, Ganesha checks if the
request's connection is part of the input session. If the check
fails, it performs the required addition of the connection to
the session. However, if the addition fails, it does not fail
the op. This is incorrect, since the client would assume that
the bind happened successfully, while the server maintained
client's session still has no knowledge of the connection. This
would cause failures in ops where Ganesha relies on the
session's connection to validate the request.

Change-Id: I5c36b5179e6cfda330178919278e597f6aad6b76
Signed-off-by: Dipit Grover <dipit@google.com>