log_functions: Don't call Cleanup from Fatal handling function
The Cleanup utility and usages (especially the uid2grp_cache_cleanup) is
not thread safe. A thread that call the LogFatal invokes now the
cleaners that destroy pthreads locks which may be used by other threads
and that may lead to undefined behavior. Also double destroy of the same
lock is undefined.
On Fatal we wand to terminate the process as fast as we can without the
risk of stuck threads that holds the process alive.

Also changing from _exit() to exit() to eliminate the ambiguity whether
the glibc version terminates just the thread or the entire process.

Signed-off-by: Assaf Yaari <assafy@google.com>
Change-Id: Ic8997399a5cc689a5bb59921a2270b7f5b347590