Reduce the default value of Readdir_Res_Size to 16KB
userspace clients like libnfs set readdirplus.maxcount to 1MB
in contrast to linux kernel clients. The kernel clients
we have experimented with typically use 16KB.

when a libnfs client tries to do a readdirplus on a huge
directory - the corresponding thread seems to acquire
directory->content_lock and doesn't release it for
> 3 minutes - this means no other client ops can
make progress during the time.

The reason we are not releasing the lock is because
we never hit tracker->mem_avail limit in
nfs3_readdir_callback which is 1MB and we keep
trying to fill more chunks

In discussions in the github, we concluded that
the best way to resolve this issue was bringing down
Readdir_Res_Size. This commit brings down
Readdir_Res_Size

Reviewers               :
Tickets resolved        :
Tickets related         :
Tests run               :
Target release          :
Breaks rolling upgrade  :
Code review URL         :

Change-Id: I7e43a98ea9dece0be308fec69db84c75ba3fb0e8
Signed-off-by: Deepak Arumugam Sankara Subramanian <deepakarumugam.s@nutanix.com>