commonlib: Place OWNER & GROUP mode generated aces at the beginning
Place the OWNER@ & GROUP@ generated aces from the mode bits at the
beginning in order to ensure that the mode-bits are taking effect so
exsiting ace with principal that happens to be the owner won't override
it. The ratio is that if the user does 'chmod', they won't be confused
by such existing ace.
Still the EVERYONE@ generated aces are placed at the end, otherwise it
will mask any other existing ace.

Signed-off-by: Assaf Yaari <assafy@google.com>
Change-Id: Ief1c0d6d07a3d76ca03d67a965272776df5cddcb