commonlib: In fsal_mode_to_acl identify if can reuse aces
When applying mode bits over NFS4 ACL, check whether existing ACL can be
reused to represent the mode.
This checks whether we have two couples of DENY & ALLOW aces at the
beginning for OWNER & GROUP, and another couple at the end for EVERYONE.
We can't rely on the existence of FSAL_ACE_IFLAG_MODE_GEN because if the
client did SETATTR for ACL, this internal flag will be lost.

Signed-off-by: Assaf Yaari <assafy@google.com>
Change-Id: Ibfa3d27b16b262516de1c86bd03d6f880df2487d