Require librgw 1.2.1
There was an issue upstream with rgw_readdir2() and versioning that can
result in ganesha failing to build against several versions of Ceph.
That API change was old enough that simplifying the requirements now for
next is a valid solution.

Just require the new version of the library and API.

Change-Id: I9d1e8d0760e17576f59c36486730fdbde6702c5d
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>