Support free disk space reporting for project quotas
FSSTAT, FSAL_VFS currently ignores the requested object
handle and calls statvfs() on the root of the export.
This means that project quotas on directories are not
considered when reporting free space. So to resolve
this change is done to get a file descriptor for the
requested object handle and call fstatvfs() on that.

This fix is similar to below FSAL change for GPFS:
299943697e4f309b01bac43742abd42f960a80f1

Change-Id: I632c6b6c37c01a5b0596c06af0f3dd4aefe5519f
Signed-off-by: Trupti Shete <tshete@ddn.com>