PROXYV3: Set KEEPALIVE and RCVTIMEO for connection between proxy & host
One interesting observation is that the read() function in proxyv3_call
will get blocked if the proxy IP becomes unavailable. In some distributed
design use cases, this can be more problematic because the proxy IP could
be moved from one node to another (causing the socket to become
unavailable, but the read() function will continue to block). It would be
helpful to monitor the socket using SO_KEEPALIVE and also set a timeout
using SO_RCVTIMEO.

Change-Id: Ia3a8182062b0808099e3f97f69c09557d678b538
Signed-off-by: Xueqian Hu <xh140312@outlook.com>