exports.c: Mount the export to the pseudo filesystem during the update process to include NFSv4.
Failed to mount using NFSv4 protocol when updating the export
protocols from version 3 to version 3,4 by DBUS command.
Additionally, Manually change the export protocols from version 3
to version 3,4, then reload the ganesha process using the SIGHUP
signal, and mounting with the NFSv4 protocol will also fail.

Both cases of mount failure mentioned above are due to the export
not being mounted to the pseudo filesystem. When the protocol of
the export is updated to include NFSv4, mounting the export to the
pseudo filesystem, which can resolve the NFSv4 mount failure issue.

The following is the relevant issue link:
https://github.com/nfs-ganesha/nfs-ganesha/issues/1128

Change-Id: Ide264c40640929ed90cba7aa980bfb013997a846
Signed-off-by: izxl007 <zengzheng007@163.com>