Fix the issue where the SIGHUP signal cannot change the export Protocols from 3 to 3,4 successfully.
Assuming there is an export with its Protocols configuration set to 3.
Then manually modify the Protocols configuration to 3,4.
After sending a SIGHUP signal to the process,
this export still cannot be mounted using the NFSv4 protocol.

The reason for this issue is that the current processing flow only handles
exports mounted on the Pseudofs root directory /.
If the export is not mounted to the Pseudofs during initialization,
it will not be mounted to the Pseudofs when the SIGHUP signal is sent.

To fix this issue, this commit modifies the handling process of the SIGHUP signal.
When the process receives a SIGHUP signal,
it will unmount all exports under the root directory and then remount them.

This solution is simple and seems quite aggressive, but it is effective.

Change-Id: Ide264c40640929ed90cba7aa980bfb013997a846
Signed-off-by: izxl007 <zengzheng007@163.com>