SAL/nfs4_recovery.c: fix the issue where adding export fails when the gethostname function call fails
Assuming the value of "sysconf(_SC_HOST_NAME_MAX)" is len.

When the actual length of the hostname exceeds (len - 1), calling
the gethostname function will result in an error, which can cause
the operation of adding new export to fail.

This patch allocates 256 bytes of space by default to store the hostname. If the length of the hostname exceeds len, it takes the
first len characters. Otherwise, it uses the hostname directly.

Change-Id: Idc2955e72a484db1d078ab4b6b2d070ab2857e89
Signed-off-by: izxl007 <zengzheng007@163.com>