Fix a bug that NFSv4 returns NFS4ERR_BAD_SEQID when the client reclaim the open states
On a high load server, there is a case that Ganasha may return
NFS4ERR_EXPIRED when it verifies the state id and find that the client
lease has expired. However, Ganasha will not free the client lease and
related resources immediately. (Ganesha use a reaper thread to do that
instead.) Once the client renews its lease and tries to reclaim the open
state, Ganesha seems to check the given seqid against the old value
saved in server. As a result, Ganasha returns NFS4ERR_BAD_SEQID.

We propose that there could be a flag to indicate whether the saved
seqid is valid or not and introduce a helper function called
Expire_nfs4_seqid to mark the seqid no longer valid every time when the
server failed to do reserve_lease. Moreover, we just find there is
already a void* pointer representing the similar meanings. We reuse the
pointer and leave comments.

Signed-off-by: BingJing Chang <bingjingc@synology.com>
Change-Id: Id5ecdb00cb6083e69d09de797d71e7d9b47403e6