Initialize read_arg->info properly for read_plus
We used to fill read_arg->info for readplus. The FSALs rely on
this buffer being filled for readplus handling. But it looks
like ac337c37db84855f406ab85c9a9583ead2a32975 deleted it.
This looks like a accident - hence reintroducing this line.

Currently VFS returns ENOTSUPP if this read_arg->info is not
NULL (i.e.) if it is a readplus operation. But since we don't
set read_arg->info properly we end up handling it like a normal
read in FSAL_VFS and finally return an empty file to the client.

Change-Id: I029c141e09b37434fd5c1801226f05dc82017408
Signed-off-by: Deepak Arumugam Sankara Subramanian <deepakarumugam.s@nutanix.com>