Default to building with prometheus monitoring enabled
In b6c59df5a0b8a3ae5b2a87446f8906cb321ee026, it was stated that
the reason for guarding monitoring behind a compile time flag
was that it introduced a new dependency. However, since
4f96c6a8e759d2a0e07643144fd003ce5c1e4e4d, there are no external
build time dependencies as a new library is used and vendored
in. This commit changes the default to include monitoring as
no external non-vendored dependencies are needed anymore.

Change-Id: I655849d330a31041cb3499de96bc6c3e17b8fbb0
Signed-off-by: YuviPanda <yuvipanda@gmail.com>