nfs4_pseudo.c: Fix the issue where subexport get disconnected from Pseudofs
Assume there are two exports:
/exp v3,v4
/exp/subexp v3,v4

Use the dbus command to change the /exp protocols from v3,v4 to v3.
As a result, /exp will be unmounted. During the pruning phase, /exp/subexp
will be unmounted first. However, it will not be remounted because /exp/subexp
does not meet the condition "export->config_gen >= generation". Eventually,
/exp/subexp will get disconnected from Pseudofs.

This patch removes the condition "export->config_gen >= generation"
to remount /exp/subexp, When /exp is to be unmounted.
This patch will not cause pseudofs structural errors in other scenarios.

Change-Id: I85edfeb199f1f52823d11dc9dd1fb84f35a4860a
Signed-off-by: izxl007 <zengzheng007@163.com>