feat(recovery_fs): reject reclaim in edge conditions
This commit enforces the state reclaim in recovery_fs backend. On
OP_RECLAIM_COMPLETE, a file named `reclaim_complete` is created
under the directory of the client id, like revoked file handles.
When recovery starts, the existence of this file fills the client
record in add_clid_entry() callback, which will be used to decide
whether the relevant client can reclaim states or not.

Change-Id: Ied2bbfd044594531a6580758032de30d16859c3e
Signed-off-by: zhitaoli <zhitao.li@iomesh.com>