Support for Preserve_Unlinked
Today, when a file is removed by an NFS client, it checks if the file is opened by other processes (or the same process). If yes, the NFS client renames the file to ".nfs<hashValue>" and when the last open is closed, removes the renamed file. This method is known as "silly rename". One problem with this approach is that the renamed files will not be removed if the client crashes.
With NFSv4.1, this logic can be moved to the server if the server supports OPEN4_RESULT_PRESERVE_UNLINKED, so a client will directly sends UNLINK for a file even it is still open. Ganesha does not support it currently and as part of this requirement we plan to implement this feature. The support on NFS client is now generally available in RHEL8.7 and RHEL9.1.

Change in this patch:
1. V4 conf param to enable/disable Preserve_Unlinked.
2. New type in fsal_fsinfo_options_t to check a FSAL's support.
3. OPEN with OPEN4_RESULT_PRESERVE_UNLINKED when supported.

Change-Id: Id763d50be454e646431bc0e38450132e685ff2ca
Signed-off-by: Xueqian Hu <xh140312@outlook.com>