add enriched bt option based on libunwind
Currently, Ganesha generates BT in the following format:
 #<line_number> <func_name> + 0x<func_offset> [ip=<hex_ip>] [sp=<hex_sp>]
The new method will generate an additional BT in the following format:
 #<line_number> at code path <code file>:<code line> from binary <binary path>

The new method is not signal safe so it's behind the build
flag USE_UNWIND_ENRICHED_BT to allow opting in or out of it.

The commit also adds a monitor thread to check if the enriched
backtrace is stuck so we can abort and complete the process exit.

Note, to enable the new backtrace you will also need
to have addr2line available during runtime

Change-Id: Iae63dc72ebcd83231dcded74c157377ba505880e
Signed-off-by: Roy Babayov <roybabayov@google.com>