ganesha-rados-grace tool: accept numeric nodeid
"ganesha-rados-grace" tool supports nodeid as text string. Using the
nodeid being passed in, it creates keys in rados object for clusterwide
grace event monitoring. The nodeid is required to be numeric (which is
part of RADOS_KV block in ganesha.conf file), and the nodename
internally formed as "node<numeric-id>", like "node1", "node5" etc.
With this change the nodeid remain numeric all over and the tool takes
care of makeing the node-name as "node<nodeid>". This will unify use of
nodeid at all places.

Change-Id: I1a648083bd24afc27b1a3f111d3f6626037c6e96
Signed-off-by: Sachin Punadikar <sachin.punadikar@ibm.com>