Config: Support IP based recovery database
The recovery database for NFS Ganesha is based on the nodeid/instance id
(based on hostname). In such scenarios, if a Ganesha instance is
listening on multiple IPs, then information about connected client
goes to a single recovery database.
In case of multiple Ganesha serving the same storage (like Ceph,
Spectrum Scale etc), then for HA functionality and load balancing, it is
beneficial to store the client information in recovery database
associated with a specific IP address to which the client is connected.
This kind of setup helps further in shifting a specific IP from one
Ganesha instance to another.
The code change is for making such confing option available.

Change-Id: I27136ed6f38142fac2c8e1dcd11c0638bbbe53b0
Signed-off-by: Sachin Punadikar <sachin.punadikar@ibm.com>