CEPH FSAL / CEPHFS: fscrypt support
In this implementation, the key is always added, and
then the "kid" that is derived from the key is used to set
the fscrypt policy.  This catches errors for when the
volume is encrypted, but with a different key.

There are no provisions in this version to remove keys;
so they will persist until ganesha is restarted.

This should build against versions of ceph that lack fscrypt
support.  In such a case, the ceph fsal will act as if all exports
are unencrypted, and reject any attempt to enable encryption.

Previous versions of this commit included linux/fscrypt.h multiple places,
but current libcephfs.h from ceph only require this if the actual fscrypt
calls are to be used.

Change-Id: I2a4d7ff38d6bcc11c969a6b21354120eddb140e1
Signed-off-by: Marcus Watts <mwatts@redhat.com>