CEPH FSAL / CEPHFS: fscrypt support
Include linux/fscrypt.h in a bunch of places.  Whether
this is actually required may depend on the version of linux,
but it never hurts.

In this implementation, the key is always added, and
then the "kid" that is derived from the key is used to set
the fscrypt policy.  This catches errors for when the
volume is encrypted, but with a different key.

There are no provisions in this version to remove keys;
so they will persist until ganesha is restarted.

ALSO NOTE: there are no provisions for to see if the relevant calls
are in ceph - so this requirse a version of ceph that includes
fscrypt support.

Change-Id: I2a4d7ff38d6bcc11c969a6b21354120eddb140e1
Signed-off-by: Marcus Watts <mwatts@redhat.com>