CONFIG: %plugin architectyure
This is a mechanism to load in code to read in custom configuration and
using that to augment the actions of fsals on a per-export basis.

Included is logic so that a plugin can request to handle configuration at a
"global" level.  After the plugin registers itself, at the appropriate
time, logic in main will call back into the plugin giving it the chance
to find and claim its configuration bits.

Also included is logic to allow custom configuration values per-export.
This uses the "CONFIG_RELAX" mechanism in config to parse the export block
several times: first time is to pick up the standard options including
the mandatory "export_id" parameter.  The second (and optional) pass
can be made by extension plugins.  Those passes must capture export_id if
they want to locate the corresponding gsh_export structure.  The last
pass accepts no parameters, it runs without CONFIG_RELAX to verify all
parameters were "found".

Also included is a "callback" mechanism so that extensions
can implement configuration changes after the root object is set.
If the callback indicates failure, the export is not made available.

Change-Id: Ibdae1aa89250a1992faaf711efd2de03264365c8
Signed-off-by: Marcus Watts <mwatts@redhat.com>