Add support for a direct SSSD usage for idmapping
This CL adds support for direct SSSD idmapping
functionality. By implementing wrappers for the
relevant functions that are exposed via
libsss_nss_idmap.so.0.

The SSSD idmap implementation is selected if
the relevant module is available and the
enum PWNAM_IMPLEMENTATION__SSSD is passed to
pwnam_wrappers__set_implementation.

This implementation allows a Ganesha to perform
direct idmapping from a SSSD server.

The implementation is built by dynamically loading
the relevant library, the implementation of the
functions is also dynamic, which make this
implementation work only on systems that have the
sssd library installed.

Change-Id: If6160c5980b9743800a172deb316ae408d294abd
Signed-off-by: Lior Suliman <liorsu@gmail.com>