fsal control: Add "control" operation fsal obj
Two operations are defined: FSCRYPT_SETKEY and
FSCRYPT_VERIFY_NOT_ENCRYPTED, both are for fscrypt support.

This is just the plumbing, it is up to individual
FSALs to implement the actual logic.  However, an important
piece of the plumbing that is included here is mdcache logic
to pass operations to the underlying FSAL.

[this was formerly:
fsal control: Add "control" operation for fsal object.
]

Change-Id: I065dd583e8ec4b6c155bc816598879f83c64a1f8
Signed-off-by: Marcus Watts <mwatts@redhat.com>