send network_id to connection manager
The network id struct is used in combination
with the client address to identify the unique client.
Using proxy protocol, it's possible for clients from
multiple networks to connect to the server, although
they have the same IP address.
This means that the client IP is not enough to identify
the client and the network id is needed as well.

Change-Id: Iff0119221344610ebc829c859d02bb5cbfe56642
Signed-off-by: Roy Babayov <roybabayov@google.com>