config: default to monitoring enabled
monitoring has moved to (lib)ntirpc, where it defaults to enabled

If we were using ntirpc only as a submodule, we could probably
eliminate this here.

But since we package ntirpc separately in Fedora, CentOS, etc.,
and in ntirpc monitoring is enabled by default, we need to keep
it here too, and have it enabled by default.

Other users who don't want monitoring can disable it in their
ntirpc (submodule or separate package) and disable it here too.

Change-Id: I9ad3574d820eaf260d7a1244488e2273165eb94d
Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>