NFS-Ganesha: Add negative-cache by UIDs
We need to be able to store the negative answers from
idmapper when mapping a uid.
This is done by introducing a new negative-cache type.

Use idmapper_negative_cache for uid when fetching groups.

If fetching groups for uid failed, meaning uid does not
exist, we would previously loop trying to get the groups
for that uid on every request.
This commit implements the negative cache, to avoid the loop.

Added uid negative lookup, and save into the cache if not found.

Change-Id: I270a7b1b64747d0cfca5513d4ed6e1649ff1038f
Signed-off-by: Lior Suliman <liorsu@gmail.com>