Remove unnecessary lookup from cih_remove_checked
If the entry is not in the avl theres no need to
do a lookup of the entry in the avl.

In my company, we try to release more entries
during reaping. From flamegraph,
the reaper thread spends a lot of
time reaping a non existent entry from the cache.
This becomes a big bottleneck when we have
to reap large number of entries. With this change
we got 30% improvement in reaping latencies.

Change-Id: Ied9b02789c5736b80dbce4dc226dd585ce6d1161
Signed-off-by: Deepak Arumugam Sankara Subramanian <deepakarumugam.s@nutanix.com>