cmake: fix conditional syntax to avoid premature variable expansion
Replaced `if(${HAS_MACRO_PREFIX_MAP_FLAG} AND USE_RELATIVE_FILE_PATHS_IN_LOGS)`
with `if(HAS_MACRO_PREFIX_MAP_FLAG AND USE_RELATIVE_FILE_PATHS_IN_LOGS)`.

Using `${...}` in `if()` can cause CMake to expand variables
before evaluating the condition, leading to errors when the
variable is unset or evaluates to `ON/OFF`. Direct variable
references avoid this issue and are the recommended style
in CMake.

Change-Id: I8188ecc73d576b390ef77f416c14039af4bfe4dd
Signed-off-by: Prabhu Murugesan <prabhu.murugesan1@ibm.com>