NFSv4.0: fix crash in release_cb_slot() due to uninitialized mutex.
In NFSv4.0, nfsv4_granted_callback() construct of RPC done by getting the chan
and basic required entities and does the RPC. But cleanup of the same is tried
w.r.t 4.1 construct. This led to a crash from accessing an uninitialized mutex.

Ensure that nfs41_release_single() is only invoked for NFSv4.1, not for NFSv4.0.

Change-Id: Id20ef65a92c2b3d1c921141ca81a6dcd67c84a2e
Signed-off-by: Deeraj.Patil <Deeraj.Patil@ibm.com>