Replace libcidr
libcidr has been used by NFS Ganesha for quite some time.
Only a handful of functions from it were used and a
whole directory of source code exists for it.
Maintenance of it gets difficult things
like clang format changes make it difficult if fixes
were made upstream (unlikely).

CIDR has been made an opaque type, because there aren't
reasons for it to be changed by other code.

These new functions are simpler and make use of
the IP parsing and conversion functions
currently in rpc_tools.c.

The second arg of cidr_contains is just a sockaddr_t because there
was no need to implement the mask on that argument.  NFS Ganesha
doesn't require that. The function has been renamed to
cidr_contains_ip.

The unused flags arg dropped from cidr_to_str().

Add test program test_cidr.c

This required some of the ip_utils.c:cidr_* functions to
be added to MainNFSD/libganesha_nfsd.ver so they can
be linked to the test program.

The plan is to move the IP-related functions
from rpc_tools to here.  We will then have the
functions in a more easily found file.

In export_mgr.c, the behaviour for DisplayExport was kept the same,
in case, someone is parsing the output or something.  One byte is
send on dbus from the old style character arrays for ip addres and
mask. Similarly, for the protocol 1 was sent for IPv4 and 2 for IPv6.
Similarly for the CIDR version.
That behaviour has been mimicked for now.  It is probably worth
cleaning that up at some point.

Change-Id: I77103440d67f1a3318f0492122e9e2a25d535aa5
Signed-off-by: Peter Schwenke <pschwenke@ddn.com>