Block signals immediately at the beginning of the main function.
Threads inherit the blocked signal mask from their creating thread. Thus, if a
thread has been created before the signals where blocked, the signals are not
blocked for that thread, meaning that when the signal is sent to the process,
the (probably default) signal handler will be called on that thread, and the
process may get terminated as a result.
Indeed, when the logging library gets initialized and the binary is compiled
with the USE_UNWIND_ENRICHED_BT macro defined, a new thread is spawned before
the signals are masked, causing the SIGHUP signal to terminate the process
instead of reloading the configuration.

Change-Id: I8158c4d977b92fc2d2d62701c702ef569500e94f
Signed-off-by: Eli Daian <daian@google.com>