Fix notify_granted_completion for NFSv4.1
This fixes an issue introduced in:
https://github.com/nfs-ganesha/nfs-ganesha/commit/1cfef0bde73e0dca0a347df7e0cad9d31ea57599

The patch changed it so up channel slot release is only called for NFSv4.1, but it used the clientid under the channel.
In NFSv4.1 the clientid is not initialized in the channel. The channel keeps a union between the clientid (for NFSv4.0) and the session (for nfsv4.1).

So, in NFSv4.1, the check looked at garbage, which means that in some cases the slot was never released.

We change the check to look at the channel type, which is always initialized.

Change-Id: I9565b148cfdd6a8d1e663e7d12f0d2705294899b
Signed-off-by: Shahar Hochma <shaharhoch@google.com>