Fix RDDR11/RDDR12 pynfs failures by ignoring dircount==0 as entry limit
READDIR was returning NFS4ERR_TOOSMALL when dircount=0,
as it was treated as “no entries allowed.” As per RFC 8881,
dircount==0 means “no limit,” and the server should rely on
maxcount to bound the reply.

This change skips the entry-limit check when dircount==0,
allowing correct NFS4_OK or NFS4ERR_ACCESS responses.

Change-Id: I12bff3c0bbfbda31d8ab9cc340cf99dfe0d63b4f
Signed-off-by: VidyaThumukunta <Vidya.Thumukunta@ibm.com>