Fix build when USE_MONITORING is not defined
A recent change introduced a dependency on <proc/readproc.h>
and `proc_t`, which causes build failures when USE_MONITORING
is not defined or procps is not available.

This commit adds a dummy definition of `proc_t` inside a
fallback `#else` block when HAVE_PROCPS is not defined,
resolving the build issue.

This is required for compiling dynamic_metrics__mem_info()
without procps support.

Change-Id: I8c7135a26b44834b9634940cd9e2f4e1f556471a
Signed-off-by: Prabhu Murugesan <prabhu.murugesan1@ibm.com>