SAL: Allow IP based recovery from RecovDir
Currently, for clustered systems the CLIDs are created in
       RecoveryRoot/RecovDir/nodeX/xxxxx

The CLIDs for recovery are read from
       RecoveryRoot/aa.bb.cc.dd
where aa,bb,cc,dd is an IP address.

This requires the node management software to copy
and delete the CLIDs.

The approach here allows the CLIDs to exists in
            RecoveryRoot/RecovDir/ip_IP_ADDRESS_STRING
where IP_ADDRESS_STRING is an address the Ganesha
instance is listening on.  The format is either IPv6
or IPv4.

The CLIDs are created in that directory and
takeover events will read from that directory.

Change-Id: Iac47a84e5c6559d37b9005e8b208bf5f4874dbbc
Signed-off-by: Peter Schwenke <pschwenke@ddn.com>