Move full debug log after NULL state check in fsal_start_io()
The debug log statement printing the open state pointer and
state type in fsal_start_io() was placed before checking if
state is NULL. This could lead to potential dereference of a
NULL pointer when full debug logging is enabled.

This patch moves the LogFullDebug() call after the NULL check
to ensure safe access to state fields.

Change-Id: I68167e3d8209221559285e49690f5b28feaab081
Signed-off-by: Suhas Athani <Suhas.Athani@ibm.com>