Fix lock leak. return lock error if we get stale lock requests of released owners
In case where lock request is received when state was already released,
We could have still served a lock. this caused a lock leak in ganesha
if lock was then added to file lock list. but if it didn't we would
have gotten a lock leak in the underlying storage

Change-Id: I27dd2bbed838bef8b6fcc088946a7ba877de8882
Signed-off-by: Ofir Vainshtein <ofirvins@google.com>